[TEST] fix read timeout
The sender waits 10 seconds before sending to make sure the receiver is
already sleeping in accept(). The receiver on the other hand has 10
seconds timeout for reading its data. Thus, in the usual setup the
receiver timeout will hit in just when the sender thinks it waited long
enough to prevent the race condition.